🛡️ Sentinel: Fix CSRF platform bypass and implement secure IDs with Edge compatibility
- Removed broad suffix-based origin matching for .vercel.app and .pages.dev in CSRF validation.
- Created `src/lib/id-utils.ts` for standardized secure ID and UUID generation using `globalThis.crypto.randomUUID()`.
- Refactored `logger.ts`, `session-analytics.ts`, `analytics.ts`, `errors.ts`, `timeline.ts`, `request-signer.ts`, and `manager.ts` to use standardized secure ID generation.
- Replaced `node:crypto` with a lightweight hash function in `errors.ts` and `request-signer.ts` for Edge Runtime (Cloudflare Workers) compatibility.
- Updated security journal with critical learnings.
