+++ src/lib/logger.ts
@@ -92,14 +92,19 @@ export function getCorrelationId(): string | undefined {
   return globalCorrelationId;
 }
 
-import { generateSecureId } from './id-utils';
-
 /**
  * Generate a unique correlation ID for request tracing
- * Uses standardized secure ID generation
+ * Uses crypto.randomUUID() for cryptographically secure, collision-resistant IDs
  */
 export function generateCorrelationId(): string {
-  return generateSecureId('req');
+  // crypto.randomUUID() is available in Node.js 15.6+ and all modern browsers
+  // Falls back to a timestamp-based ID if crypto is not available (rare edge case)
+  try {
+    return `req_${crypto.randomUUID()}`;
+  } catch {
+    // Fallback for environments without crypto.randomUUID support
+    return `req_${Date.now()}_${Math.random().toString(36).substring(2, 11)}`;
+  }
 }
 
 export interface LogContext {
