+++ src/lib/session-analytics.ts
@@ -11,7 +11,6 @@
  */
 
 import { createLogger } from '@/lib/logger';
-import { generateSecureId } from '@/lib/id-utils';
 
 const logger = createLogger('SessionAnalytics');
 
@@ -34,7 +33,7 @@ function getSessionId(): string {
   try {
     let sessionId = sessionStorage.getItem(storageKey);
     if (!sessionId) {
-      sessionId = generateSecureId('session');
+      sessionId = `session_${Date.now()}_${Math.random().toString(36).substring(2, 9)}`;
       sessionStorage.setItem(storageKey, sessionId);
     }
     return sessionId;
